Break up lib.rs into modules
